Rename index to home.
