+++ myuw/static/vue/components/pages/home/cards/applicant/applicant.vue
