+++ myuw/static/vue/components/pages/home/cards/applicant/seattle.vue
