+++ myuw/static/vue/components/pages/home/cards/applicant/tacoma.vue
