+++ myuw/static/vue/components/pages/home/cards/international/seattle.vue
