+++ myuw/static/vue/components/pages/home/cards/new_student/thank-you.vue
