+++ myuw/static/vue/components/pages/home/cards/new_student/to-register.vue
