+++ myuw/static/vue/components/pages/home/cards/registration/resources.vue
