+++ myuw/static/vue/components/pages/home/cards/registration/status.vue
