+++ myuw/static/vue/components/pages/home/summaries.vue
